# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rhytmiq API",
    "description": "FIG-compliant API for managing rhythmic gymnastics meets",
    "version": "0.1.0"
  },
  "paths": {
...
    "/meets/": {
      "post": {
        "tags": [
          "Meets"
        ],
        "summary": "Create Meet",
        "operationId": "create_meet_meets__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Meets"
        ],
        "summary": "List Meets",
        "operationId": "list_meets_meets__get",
        "parameters": [
          {
            "name": "district_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by district_id",
              "title": "District Id"
            },
            "description": "Filter by district_id"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/MeetStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status",
              "title": "Status"
            },
            "description": "Filter by status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeetRead"
                  },
                  "title": "Response List Meets Meets  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/meets/{meet_id}/standings": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Get Apparatus Standings",
        "operationId": "get_apparatus_standings_meets__meet_id__standings_get",
        "parameters": [
          {
            "name": "meet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meet Id"
            }
          },
          {
            "name": "apparatus",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Apparatus",
              "description": "Apparatus to rank (required)."
            },
            "description": "Apparatus to rank (required)."
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Level"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by level",
              "title": "Level"
            },
            "description": "Filter by level"
          },
          {
            "name": "age_group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AgeGroup"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by age_group",
              "title": "Age Group"
            },
            "description": "Filter by age_group"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApparatusStandingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/meets/{meet_id}/all-around": {
      "get": {
        "tags": [
          "Results"
        ],
        "summary": "Get All Around Standings",
        "operationId": "get_all_around_standings_meets__meet_id__all_around_get",
        "parameters": [
          {
            "name": "meet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Meet Id"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Level"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by level",
              "title": "Level"
            },
            "description": "Filter by level"
          },
          {
            "name": "age_group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AgeGroup"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by age_group",
              "title": "Age Group"
            },
            "description": "Filter by age_group"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllAroundStandingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgeGroup": {
        "type": "string",
        "enum": [
          "u7",
          "u8",
          "u9",
          "u10",
          "u11",
          "o11",
          "u12",
          "u14",
          "o14"
        ],
        "title": "AgeGroup"
      },
      "AllAroundStandingRow": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank",
            "description": "Competition rank within this all-around category."
          },
          "entry_id": {
            "type": "integer",
            "title": "Entry Id",
            "description": "The meet entry this row represents."
          },
          "competitor_name": {
            "type": "string",
            "title": "Competitor Name",
            "description": "Gymnast full name, or group name."
          },
          "bib_number": {
            "type": "string",
            "title": "Bib Number",
            "description": "Bib number of the competing entry."
          },
          "level": {
            "$ref": "#/components/schemas/Level"
          },
          "age_group": {
            "$ref": "#/components/schemas/AgeGroup"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total",
            "description": "Sum of this entry's routine totals."
          },
          "e_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "E Total",
            "description": "Sum of this entry's routine Execution scores."
          },
          "routines_counted": {
            "type": "integer",
            "title": "Routines Counted",
            "description": "Number of routines summed -- less than the full apparatus count means a partial (in-progress or incomplete) all-around."
          },
          "medal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "gold",
                  "silver",
                  "bronze"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Medal",
            "description": "Standard-based medal tier from the meet's configured cutoffs (Meet.medal_gold_min/medal_silver_min), independent of rank. Null if the meet isn't using cutoffs."
          }
        },
        "type": "object",
        "required": [
          "rank",
          "entry_id",
          "competitor_name",
          "bib_number",
          "level",
          "age_group",
          "total",
          "e_total",
          "routines_counted"
        ],
        "title": "AllAroundStandingRow"
      },
      "AllAroundStandingsRead": {
        "properties": {
          "meet_id": {
            "type": "integer",
            "title": "Meet Id"
          },
          "provisional": {
            "type": "boolean",
            "title": "Provisional",
            "description": "True unless the meet's status is 'completed'."
          },
          "level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Level"
              },
              {
                "type": "null"
              }
            ],
            "description": "Level filter applied, if any."
          },
          "age_group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgeGroup"
              },
              {
                "type": "null"
              }
            ],
            "description": "Age group filter applied, if any."
          },
          "rankings": {
            "items": {
              "$ref": "#/components/schemas/AllAroundStandingRow"
            },
            "type": "array",
            "title": "Rankings"
          }
        },
        "type": "object",
        "required": [
          "meet_id",
          "provisional",
          "rankings"
        ],
        "title": "AllAroundStandingsRead",
        "description": "Live-computed all-around ranking for one meet/level/age_group slice."
      },
      "Apparatus": {
        "type": "string",
        "enum": [
          "rope",
          "hoop",
          "ball",
          "clubs",
          "ribbon",
          "freehand"
        ],
        "title": "Apparatus"
      },
      "ApparatusStandingRow": {
        "properties": {
          "rank": {
            "type": "integer",
            "title": "Rank",
            "description": "Competition rank within this apparatus category."
          },
          "entry_id": {
            "type": "integer",
            "title": "Entry Id",
            "description": "The meet entry this routine belongs to."
          },
          "routine_id": {
            "type": "integer",
            "title": "Routine Id",
            "description": "The routine this row represents."
          },
          "competitor_name": {
            "type": "string",
            "title": "Competitor Name",
            "description": "Gymnast full name, or group name."
          },
          "bib_number": {
            "type": "string",
            "title": "Bib Number",
            "description": "Bib number of the competing entry."
          },
          "level": {
            "$ref": "#/components/schemas/Level"
          },
          "age_group": {
            "$ref": "#/components/schemas/AgeGroup"
          },
          "apparatus": {
            "$ref": "#/components/schemas/Apparatus"
          },
          "d_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "D Score"
          },
          "a_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "A Score"
          },
          "e_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "E Score"
          },
          "final_score": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Final Score",
            "description": "Levels 1-3 only: the single pre-aggregated mark out of 13. 0 at every other level, where the score is built from the D/A/E panels instead."
          },
          "penalty": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Penalty"
          },
          "total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total"
          },
          "medal": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "gold",
                  "silver",
                  "bronze"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Medal",
            "description": "Standard-based medal tier from the meet's configured cutoffs (Meet.medal_gold_min/medal_silver_min), independent of rank. Null if the meet isn't using cutoffs."
          }
        },
        "type": "object",
        "required": [
          "rank",
          "entry_id",
          "routine_id",
          "competitor_name",
          "bib_number",
          "level",
          "age_group",
          "apparatus",
          "d_score",
          "a_score",
          "e_score",
          "final_score",
          "penalty",
          "total"
        ],
        "title": "ApparatusStandingRow"
      },
      "ApparatusStandingsRead": {
        "properties": {
          "meet_id": {
            "type": "integer",
            "title": "Meet Id"
          },
          "provisional": {
            "type": "boolean",
            "title": "Provisional",
            "description": "True unless the meet's status is 'completed'."
          },
          "apparatus": {
            "$ref": "#/components/schemas/Apparatus"
          },
          "level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Level"
              },
              {
                "type": "null"
              }
            ],
            "description": "Level filter applied, if any."
          },
          "age_group": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgeGroup"
              },
              {
                "type": "null"
              }
            ],
            "description": "Age group filter applied, if any."
          },
          "rankings": {
            "items": {
              "$ref": "#/components/schemas/ApparatusStandingRow"
            },
            "type": "array",
            "title": "Rankings"
          }
        },
        "type": "object",
        "required": [
          "meet_id",
          "provisional",
          "apparatus",
          "rankings"
        ],
        "title": "ApparatusStandingsRead",
        "description": "Live-computed per-apparatus ranking for one meet/level/age_group/apparatus slice."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Level": {
        "type": "string",
        "enum": [
          "level_1",
          "level_2",
          "level_3",
          "level_4",
          "level_5",
          "level_6",
          "level_7",
          "level_8",
          "level_9",
          "level_10",
          "high_performance_1",
          "high_performance_2",
          "high_performance_3",
          "high_performance_4",
          "pre_junior",
          "junior",
          "senior",
          "olympic"
        ],
        "title": "Level"
      },
      "MeetCreate": {
        "properties": {
          "district_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name"
          },
          "location": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Location"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "status": {
            "$ref": "#/components/schemas/MeetStatus",
            "default": "draft"
          },
          "medal_gold_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medal Gold Min"
          },
          "medal_silver_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medal Silver Min"
          }
        },
        "type": "object",
        "required": [
          "name",
          "location",
          "start_date",
          "end_date"
        ],
        "title": "MeetCreate"
      },
...
      "MeetRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "status": {
            "$ref": "#/components/schemas/MeetStatus"
          },
          "medal_gold_min": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medal Gold Min"
          },
          "medal_silver_min": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medal Silver Min"
          }
        },
        "type": "object",
        "required": [
          "id",
          "district_id",
          "name",
          "location",
          "start_date",
          "end_date",
          "status",
          "medal_gold_min",
          "medal_silver_min"
        ],
        "title": "MeetRead"
      },
      "MeetStatus": {
        "type": "string",
        "enum": [
          "draft",
          "scheduled",
          "in_progress",
          "completed",
          "cancelled"
        ],
        "title": "MeetStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}